stem_agent/pyproject.toml
[project]
name = "stem-agent"
version = "0.1.0"
description = "A stem agent that grows into a specialized Deep Research agent via self-differentiation"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
search = [
    "tavily-python>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stem_agent"]
